requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
chromadb
google-genai
pydantic
sqlalchemy
sentence-transformers
